add some simple seminar operation

+++ src/main/java/com/loha/flippedclassroom/mapper/TeamMapper.java
@@ -28,5 +28,5 @@ public interface TeamMapper {
      * @return team list
      * @throws Exception
      */
-    List<Team> selectTeamByCourseId(Integer courseId) throws Exception;
+    List<Team> selectTeamByCourseId(long courseId) throws Exception;
 }
